Add ForthEngine and ForthWord functions to 'forth' program
This commit creates a new 'ForthEngine' interface and extends 'ForthWord' functions such as addition, subtraction, and multiplication in the 'ForthStack' for the 'forth' program.

'ForthEngine' carries the 'ForthStack' interface and also integrates the consumer functional interface to allow the inner workings of the forth program to modify the stack according to received commands.

'ForthWord' now includes representation for the main arithmetic operations, as well as functions to test for number values and duplications. Furthermore, built-in checks have been implemented in these methods to ensure sufficient values are available in the stack for each operation, enhancing overall program reliability.

Lastly, 'ForthCore' now serves as the main operator in the forth program, compiling tokens and ensuring word validation. This structural change allows better encapsulation and code maintainability in larger scale developments.