Add BinaryConverter class to encode strings into ASCII binary
Introduced a new BinaryConverter class located in the 'com.epam.chucknorris.encode' package. This class provides functionality to convert strings into binary representations using ASCII encoding. It implements the UnaryOperator interface, which allows it to be used as a function via the apply method. This change facilitates efficient string-to-binary conversion in the context of internal processing and/or external communications.