Implement symmetric tree check for LeetCode problem 101
This commit introduces added functionality to check if a binary tree is symmetric or not. A new file, Solution.java contains the primary method isSymmetric() that compares the left and right sub-trees of the root. Other supporting function isSymmetric(TreeNode leftNode, TreeNode rightNode) recursively aids in the comparison.

In addition, two versions of the TreeNode class (with slight variations) were also included in the commit. Both include constructors and definitions required for working with binary trees in the context of this problem. One of the versions (in leetcode package) includes a toString() method to represent the TreeNode in a specific format.