Add test cases and solution for LeetCode problem 73
Implemented the solution for LeetCode problem 73 - Set Matrix Zeroes, where if an element in matrix is 0, its entire row and column is set to 0. Created two test classes with various test cases to validate the functionality of the solution. The test cases involve matrices of various sizes and configurations. This solution uses a BitSet for tracking rows and columns contain zero in the matrix and has time complexity of O(m*n).