Add SmallestValueFunction and related tests
This commit introduces the SmallestValueFunction and its related test cases. The function is meant to solve the problem of finding the smallest integer n that satisfies n! > m for a given long positive number m. Different test cases illustrating various scenarios have been added. The user instruction on using the function is also included in 'docs/the-smallest-value.md'.