Refactor maze generation logic for flexibility and scalability
Reformed the maze generation logic in MazeGenerator to provide more flexible and scalable solutions. The previous static maze generation has been replaced with a dynamic method using depth-first-search algorithm which enables us to create mazes of varying sizes. new Grid and Direction classes were introduced to aid in this. BitSet is used for grid representation resulting in better performance and reduced memory use. This provides the application with the ability to generate a wide variety of mazes based on user-input parameters.