Refactor maze generation for testablility and determinism
The changes add a `randomGenerator` parameter to the `Grid` constructor to allow a deterministic maze generation by controlling the seed. The `MazeGenerator` can now generate perfect mazes with and without seeds to provide testable conditions. Additionally, a `MazeChecker` test module is created to check the generated maze's correctness. Several tests are added to verify the mazes generated with and without seeds and to test small square mazes. The file `Maze.java` was renamed to `MazeParameters.java` for clarity and the file `MazeReport.java` was deleted as it's no longer needed For specified mazes, function `generatePerfectMaze` now supports an optional seed parameter. Additionally, a 'MazeChecker' class was added in the test package to verify the correctness of the generated maze.