Change Grid's symbol() method to return char
The Grid's symbol() method has been updated to return a primitive char type instead of Character object for improving the performance of the maze generating algorithm. This change helps to avoid unnecessary autoboxing and unboxing of chars that happens when converting a char into a Character and vice versa, thereby reducing the overhead of object creation.