Add test for single exit validation in 'MazeGeneratorTest'
This commit introduces a new test titled 'theMazeHasASingleExitOnTheRightSideOfTheMaze' to 'MazeGeneratorTest'. The aim of this test is to confirm that the maze generated only comprises a single exit on the right side. A helper method, 'countExits', was also implemented in order to facilitate this test. This ensures stricter adherence to the maze design requirements, thus improving game consistency.

closes #56