Add solution for leetcode problem 338 and corresponding tests
Implemented the 'countBits' method in the Solution.java file for leetcode problem number 338. This method uses the Java 'Integer.bitCount(n)' function to count the number of 1s in the binary representation of a number up to 'n' and return an array of these counts.

Additionally, a test class named 'SolutionBlackBoxC4Test' was added which tests various scenarios including edge cases to ensure the functionality and reliability of the 'countBits' method.