Add solution for leetcode problem 268 and corresponding tests
The solution for Leetcode problem 268 has been implemented in a new file 'Solution.java'. The solution consists of 'missingNumber' method, which calculates the expected sum of numbers in an array, subtracts it with the actual sum of numbers to find the missing one.

Associated unit tests have been written in 'SolutionBlackBoxC4Test.java'. Tests include various scenarios with different array sizes and order to ensure the functionality and robustness of the solution.